update BaseTest helper methods to use longs. update OrderTest. update MarketOrderTest.

+++ src/test/java/org/anhonesteffort/btc/book/BaseTest.java
@@ -20,37 +20,37 @@ package org.anhonesteffort.btc.book;
 public class BaseTest {
 
   private Integer nextOrderId = 0;
-  private long serial = 0;
+  private Long serial = 0l;
 
-  protected Order newAsk(String orderId, double price, double size) {
+  protected Order newAsk(String orderId, long price, long size) {
     return new Order(serial++, orderId, Order.Side.ASK, price, size);
   }
 
-  protected Order newAsk(double price, double size) {
+  protected Order newAsk(long price, long size) {
     return newAsk((nextOrderId++).toString(), price, size);
   }
 
-  protected MarketOrder newMarketAsk(String orderId, double size, double funds) {
+  protected MarketOrder newMarketAsk(String orderId, long size, long funds) {
     return new MarketOrder(serial++, orderId, Order.Side.ASK, size, funds);
   }
 
-  protected MarketOrder newMarketAsk(double size, double funds) {
+  protected MarketOrder newMarketAsk(long size, long funds) {
     return newMarketAsk((nextOrderId++).toString(), size, funds);
   }
 
-  protected Order newBid(String orderId, double price, double size) {
+  protected Order newBid(String orderId, long price, long size) {
     return new Order(serial++, orderId, Order.Side.BID, price, size);
   }
 
-  protected Order newBid(double price, double size) {
+  protected Order newBid(long price, long size) {
     return newBid((nextOrderId++).toString(), price, size);
   }
 
-  protected MarketOrder newMarketBid(String orderId, double size, double funds) {
+  protected MarketOrder newMarketBid(String orderId, long size, long funds) {
     return new MarketOrder(serial++, orderId, Order.Side.BID, size, funds);
   }
 
-  protected MarketOrder newMarketBid(double size, double funds) {
+  protected MarketOrder newMarketBid(long size, long funds) {
     return newMarketBid((nextOrderId++).toString(), size, funds);
   }
